+++ extensions/whois-tool/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fix for "Extensions / Check"
 
+## [fix] - 2024-08-18
+
+- Fix for "Extensions / Check"
+
 ## [whois-tool] - 2024-08-18
 
 - Add the main code for the Whois-tool extension